requirements.txt
altair==5.2.0
attrs==23.2.0
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
filelock==3.13.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
openai-whisper==20231117
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
streamlit==1.31.0
streamlit-audiorec==0.1.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
